add reservation form component
